requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
Flask-Cors==3.0.10
openai-whisper==20231117
deep-translator==1.11.4
gTTS==2.3.2
pydub==0.25.1
waitress==2.1.2